requirements_cpu.txt
pydantic
python-dotenv==1.0.1
fastapi
uvicorn
python-multipart
httptools
openvino
opencv-python
scikit-learn
optimum[openvino]
transformers
pillow
numpy
requests
